+++ src/main/java/wooteco/subway/config/WebMvcConfig.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import wooteco.subway.web.member.LoginMemberMethodArgumentResolver;
@@ -15,11 +14,6 @@ import wooteco.subway.web.member.interceptor.BearerAuthInterceptor;
 public class WebMvcConfig implements WebMvcConfigurer {
     private final BearerAuthInterceptor bearerAuthInterceptor;
     private final LoginMemberMethodArgumentResolver loginMemberArgumentResolver;
-
-    @Override
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/admin").setViewName("admin/index");
-    }
 
     public WebMvcConfig(BearerAuthInterceptor bearerAuthInterceptor,
         LoginMemberMethodArgumentResolver loginMemberArgumentResolver) {
