+++ src/main/java/wooteco/subway/domain/line/LineStations.java
@@ -3,7 +3,6 @@ package wooteco.subway.domain.line;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
