+++ src/main/java/wooteco/subway/web/AdminPageController.java
@@ -1,0 +1,47 @@
+package wooteco.subway.web;
+
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import wooteco.subway.service.line.LineService;
+import wooteco.subway.service.station.StationService;
+
+@Controller
+public class AdminPageController {
+    private LineService lineService;
+    private StationService stationService;
+
+    public AdminPageController(LineService lineService, StationService stationService) {
+        this.lineService = lineService;
+        this.stationService = stationService;
+    }
+
+    @GetMapping(value = "/admin", produces = MediaType.TEXT_HTML_VALUE)
+    public String admin() {
+        return "admin/index";
+    }
+
+    @GetMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
+    public String index() {
+        return "service/index";
+    }
+
+    @GetMapping(value = "/stations", produces = MediaType.TEXT_HTML_VALUE)
+    public String stationPage(Model model) {
+        model.addAttribute("stations", stationService.findStations());
+        return "admin/admin-station";
+    }
+
+    @GetMapping(value = "/lines", produces = MediaType.TEXT_HTML_VALUE)
+    public String linePage(Model model) {
+        model.addAttribute("lines", lineService.findLines());
+        return "admin/admin-line";
+    }
+
+    @GetMapping(value = "/edges", produces = MediaType.TEXT_HTML_VALUE)
+    public String edgePage() {
+        return "admin/admin-edge";
+    }
+}
