+++ src/test/java/wooteco/subway/domain/station/StationRepositoryTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.data.jdbc.repository.config.EnableJdbcAuditing;
 import org.springframework.data.relational.core.conversion.DbActionExecutionException;
 
 import wooteco.subway.config.EnableJdbcAuditingConfig;
